Consolidate ShortcodeProcessor cases into one test class
